refactor: move Byteutils to internals
