+++ clients/src/main/java/org/apache/kafka/common/record/internal/DefaultRecord.java
@@ -20,7 +20,7 @@ import org.apache.kafka.common.InvalidRecordException;
 import org.apache.kafka.common.header.Header;
 import org.apache.kafka.common.header.internals.RecordHeader;
 import org.apache.kafka.common.record.TimestampType;
-import org.apache.kafka.common.utils.ByteUtils;
+import org.apache.kafka.common.utils.internals.ByteUtils;
 import org.apache.kafka.common.utils.Utils;
 
 import java.io.DataOutputStream;
