+++ clients/src/test/java/org/apache/kafka/common/utils/ByteUtilsTest.java
@@ -16,6 +16,7 @@
  */
 package org.apache.kafka.common.utils;
 
+import org.apache.kafka.common.utils.internals.ByteUtils;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
