+++ raft/src/main/java/org/apache/kafka/raft/internals/BatchBuilder.java
@@ -27,7 +27,7 @@ import org.apache.kafka.common.record.internal.DefaultRecordBatch;
 import org.apache.kafka.common.record.internal.MemoryRecords;
 import org.apache.kafka.common.record.internal.RecordBatch;
 import org.apache.kafka.common.utils.ByteBufferOutputStream;
-import org.apache.kafka.common.utils.ByteUtils;
+import org.apache.kafka.common.utils.internals.ByteUtils;
 import org.apache.kafka.server.common.serialization.RecordSerde;
 
 import java.io.DataOutputStream;
