+++ raft/src/test/java/org/apache/kafka/raft/internals/BatchAccumulatorTest.java
@@ -28,7 +28,7 @@ import org.apache.kafka.common.record.internal.ControlRecordUtils;
 import org.apache.kafka.common.record.internal.DefaultRecord;
 import org.apache.kafka.common.record.internal.MemoryRecordsBuilder;
 import org.apache.kafka.common.record.internal.RecordBatch;
-import org.apache.kafka.common.utils.ByteUtils;
+import org.apache.kafka.common.utils.internals.ByteUtils;
 import org.apache.kafka.common.utils.MockTime;
 import org.apache.kafka.common.utils.Utils;
